add payment UI
