+++ app/profile/components/order/order-history.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import styles from './order-history.module.css';
 import { IOrder, IOrderHistory } from '@/types/interfaces/profile';
-import { options, orders as mockOrders } from '@/app/profile/[userId]/helpers';
+import { options, orders as mockOrders } from '@/app/profile/[userId]/mock-data';
 import {
   containerStyle,
   controlStyles,
