+++ app/profile/components/payment/components/payment-form.module.css
@@ -10,6 +10,18 @@
   justify-content: space-between;
   color: var(--black);
   align-items: center;
+  margin-bottom: 10px;
+}
+
+.title {
+  font-family: PoppinsS;
+  font-size: 18px;
+  font-weight: 500;
+  text-transform: uppercase;
+}
+
+.typeCardImg {
+  display: flex;
 }
 
 .input {
@@ -26,7 +38,7 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  margin-top: 36px;
+  margin-top: 32px;
 }
 
 .inputContainer {
