+++ app/profile/components/payment/payment.module.css
@@ -1,0 +1,24 @@
+.deliveryContainer {
+  max-width: 464px;
+  height: 100%;
+}
+
+.title {
+  font-family: PoppinsS;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  text-transform: uppercase;
+  color: var(--black);
+  margin-bottom: 22px;
+  padding-left: 8px;
+}
+
+.img {
+  margin: 0;
+}
+
+.img:nth-child(1) {
+  margin-right: 8px;
+}
