+++ types/interfaces/profile.ts
@@ -1,5 +1,5 @@
 import { CSSProperties } from 'react';
-import { ActionMeta, CSSObjectWithLabel, SingleValue } from 'react-select';
+import { SingleValue } from 'react-select';
 
 export interface IProfileItem {
   id: number;
@@ -30,31 +30,6 @@ export interface IOrder {
   phone: string;
 }
 
-export interface IOrderHistory {}
-
-export interface ISelect {
-  options: IOptions[];
-  onChange: (newValue: SingleValue<IOptions>) => void;
-  defaultOption?: IOptions;
-  placeholder?: string;
-  indicatorSeparatorStyle?: CSSProperties;
-  controlStyle?: CSSProperties;
-  containerStyle?: CSSProperties;
-  indicatorContainerStyle?: CSSProperties;
-  dropdownIndicatorStyle?: CSSProperties;
-  menuStyle?: CSSProperties;
-  optionStyle?: CSSProperties;
-  selectedColor?: string;
-  hideSelectedOptions?: boolean;
-  isSearchable?: boolean;
-}
-
-export interface IProfileOrderTable {
-  orders: IOrder[];
-  handleRowClick: (id: number) => void;
-  selectedRow: number | null;
-}
-
 export interface IReceipt {
   image: string;
   name: string;
@@ -65,7 +40,7 @@ export interface IReceipt {
   total: string;
 }
 
-export interface IForm {
+export interface IFormDelivery {
   country?: string;
   city?: string;
   region?: string;
@@ -73,12 +48,4 @@ export interface IForm {
   building?: string;
   unit?: string;
   phoneNumber?: string;
-}
-
-export interface IAddressForm {
-  form: IForm;
-  index: number;
-  handleInputChange: (index: number, field: keyof IForm, value: string) => void;
-  removeForm: (index: number) => void;
-  isFormFilled: (form: IForm) => boolean;
 }
